refactor!: adopt the sql.dialect coordinates from the jdbc.db split
The dialect framework moved to the org.eclipse.daanse.sql repository
(plain file copies; history in jdbc.db). Mechanical rename of imports
and pom artifactIds: org.eclipse.daanse.jdbc.db.dialect.* ->
org.eclipse.daanse.sql.dialect.*.
