+++ mariadb/test.bndrun
@@ -66,7 +66,6 @@
 	org.eclipse.daanse.jdbc.datasource.mariadb;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.jdbc.datasource.mariadb-tests;version='[0.0.1,0.0.2)',\
 	org.opentest4j;version='[1.3.0,1.3.1)',\
-	org.osgi.service.cm;version='[1.6.1,1.6.2)',\
 	org.osgi.service.component;version='[1.5.1,1.5.2)',\
 	org.osgi.test.common;version='[1.3.0,1.3.1)',\
 	org.osgi.test.junit5;version='[1.3.0,1.3.1)',\
