+++ mssqlserver/test.bndrun
@@ -52,7 +52,7 @@
 # -runbundles is calculated by the bnd-resolver-maven-plugin
 -runbundles: \
 	assertj-core;version='[3.26.0,3.26.1)',\
-	com.microsoft.sqlserver.mssql-jdbc;version='[12.8.1,12.8.2)',\
+	com.microsoft.sqlserver.mssql-jdbc;version='[12.8.2,12.8.3)',\
 	junit-jupiter-api;version='[5.12.2,5.12.3)',\
 	junit-jupiter-engine;version='[5.12.2,5.12.3)',\
 	junit-jupiter-params;version='[5.12.2,5.12.3)',\
@@ -66,7 +66,6 @@
 	org.eclipse.daanse.jdbc.datasource.mssqlserver;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.jdbc.datasource.mssqlserver-tests;version='[0.0.1,0.0.2)',\
 	org.opentest4j;version='[1.3.0,1.3.1)',\
-	org.osgi.service.cm;version='[1.6.1,1.6.2)',\
 	org.osgi.service.component;version='[1.5.1,1.5.2)',\
 	org.osgi.service.jdbc;version='[1.1.0,1.1.1)',\
 	org.osgi.test.common;version='[1.3.0,1.3.1)',\
