+++ testkit/api/src/main/java/org/eclipse/daanse/jdbc/datasource/testkit/api/DatabaseProvider.java
@@ -16,7 +16,7 @@ import java.util.ServiceLoader;
 
 import javax.sql.DataSource;
 
-import org.eclipse.daanse.jdbc.db.dialect.api.Dialect;
+import org.eclipse.daanse.sql.dialect.api.Dialect;
 
 /**
  * SPI: provides a {@link DataSource} + matching {@link Dialect}. Discovered via
