+++ testkit/mariadb/pom.xml
@@ -35,7 +35,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
-      <artifactId>org.eclipse.daanse.jdbc.db.dialect.db.mariadb</artifactId>
+      <artifactId>org.eclipse.daanse.sql.dialect.db.mariadb</artifactId>
       <version>${revision}</version>
       <scope>compile</scope>
     </dependency>
