+++ testkit/mssqlserver/pom.xml
@@ -35,7 +35,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
-      <artifactId>org.eclipse.daanse.jdbc.db.dialect.db.mssqlserver</artifactId>
+      <artifactId>org.eclipse.daanse.sql.dialect.db.mssqlserver</artifactId>
       <version>${revision}</version>
       <scope>compile</scope>
     </dependency>
