+++ testkit/oracle/pom.xml
@@ -33,7 +33,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
-      <artifactId>org.eclipse.daanse.jdbc.db.dialect.db.oracle</artifactId>
+      <artifactId>org.eclipse.daanse.sql.dialect.db.oracle</artifactId>
       <version>${revision}</version>
       <scope>compile</scope>
     </dependency>
