+++ testkit/postgresql/src/main/java/org/eclipse/daanse/jdbc/datasource/testkit/postgresql/PostgresDatabaseProvider.java
@@ -20,9 +20,9 @@ import java.util.concurrent.ConcurrentMap;
 
 import org.eclipse.daanse.jdbc.datasource.testkit.api.ActiveDatabase;
 import org.eclipse.daanse.jdbc.datasource.testkit.api.DatabaseProvider;
-import org.eclipse.daanse.jdbc.db.dialect.api.Dialect;
-import org.eclipse.daanse.jdbc.db.dialect.api.DialectInitData;
-import org.eclipse.daanse.jdbc.db.dialect.db.postgresql.PostgreSqlDialect;
+import org.eclipse.daanse.sql.dialect.api.Dialect;
+import org.eclipse.daanse.sql.dialect.api.DialectInitData;
+import org.eclipse.daanse.sql.dialect.db.postgresql.PostgreSqlDialect;
 import org.postgresql.ds.PGSimpleDataSource;
 import org.testcontainers.containers.PostgreSQLContainer;
