+++ testkit/sqlite/pom.xml
@@ -36,7 +36,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
-      <artifactId>org.eclipse.daanse.jdbc.db.dialect.db.sqlite</artifactId>
+      <artifactId>org.eclipse.daanse.sql.dialect.db.sqlite</artifactId>
       <version>${revision}</version>
       <scope>compile</scope>
     </dependency>
