+++ testkit/sqlite/src/main/java/org/eclipse/daanse/jdbc/datasource/testkit/sqlite/SqliteDatabaseProvider.java
@@ -20,9 +20,9 @@ import java.util.concurrent.ConcurrentMap;
 
 import org.eclipse.daanse.jdbc.datasource.testkit.api.ActiveDatabase;
 import org.eclipse.daanse.jdbc.datasource.testkit.api.DatabaseProvider;
-import org.eclipse.daanse.jdbc.db.dialect.api.Dialect;
-import org.eclipse.daanse.jdbc.db.dialect.api.DialectInitData;
-import org.eclipse.daanse.jdbc.db.dialect.db.sqlite.SqliteDialect;
+import org.eclipse.daanse.sql.dialect.api.Dialect;
+import org.eclipse.daanse.sql.dialect.api.DialectInitData;
+import org.eclipse.daanse.sql.dialect.db.sqlite.SqliteDialect;
 import org.sqlite.SQLiteDataSource;
 
 /**
